% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_representation.R
\name{adj_CDF_logit}
\alias{adj_CDF_logit}
\title{A novel data representation based on Poisson probability}
\usage{
adj_CDF_logit(test_set, change = 1e-10)
}
\arguments{
\item{change}{a numeric value used to correct for exactly 0 and 1 before logit transformation.
Any values below \code{change} are set to be \code{change} and
any values above \eqn{1- \code{change}} are set to be \eqn{1- \code{change}}.}

\item{dat}{a UMI count data matirx with cells as rows and genes as columns}
}
\value{
a matrix of departure as a novel data representation.
}
\description{
adj_CDF_logit returns a matrix of novel data representation with the same dimension as input data matrix.
}
\details{
This is a function used to calcualte model departure as a novel data representation.
}
\examples{

test_set <- matrix(rpois(500, 0.5), nrow = 10)
para_est_new(test_set)

}
