% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clust_clean}
\alias{clust_clean}
\title{Cluster label clean}
\usage{
clust_clean(clust)
}
\arguments{
\item{clust}{a string indicates cluster label at each split step}
}
\value{
a string with unwanted characters removed
}
\description{
This function removes unwanted characters from cluster label string
}
\details{
The clust_clean function removes any "-" or "NA" at the end of a string for a given cluster label
}
