% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cluster_size}
\alias{cluster_size}
\title{Cluster size}
\usage{
cluster_size(test_dat)
}
\arguments{
\item{test_dat}{a matrix or data frame with cells to cluster as rows}
}
\value{
a numeric value with number of cells to cluster
}
\description{
This function calculates the number of elements in current cluster
}
