% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{diff_gene_list.matrix2}
\alias{diff_gene_list.matrix2}
\title{Update departure matrix based on clusters used for comparison (input should be the raw UMI count matrix)}
\usage{
\method{diff_gene_list}{matrix2}(data, final_clust_res, clust1 = "1", clust2 = "2", t = F)
}
\value{
scppp_de_results
}
\description{
Update departure matrix based on clusters used for comparison (input should be the raw UMI count matrix)
}
