% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{get_example_data}
\alias{get_example_data}
\title{get example data}
\usage{
get_example_data(x = c("p5", "p56"))
}
\arguments{
\item{x}{data set to choose}
}
\value{
A data set from example data
}
\description{
get example data
}
