% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qq_env_small.R
\name{interpolate}
\alias{interpolate}
\title{Linear interpolation for one sample given reference sample}
\usage{
interpolate(df, reference, sample_id)
}
\arguments{
\item{df}{The object data frame need to do interpolation}

\item{reference}{The reference data frame to make comparison}

\item{sample_id}{A character to denote object data frame}
}
\value{
A data frame. First column is CDF, second column is sample name,
and third column is correponsding quantiles.
}
\description{
interpolate returns a data frame with interpolated data points
}
\details{
This is a function developed to do linear interpolation for corresponding probability
from empirical cumulative distribution function (CDF) and corresponding quanitles.
Given a reference data frame and a data frame needed to do interpolation,
if there are any CDF values in reference but not in object data frame,
do the linear interpolation and insert both CDF values and respective quantiles
to the original object data frame.
}
\examples{

P <- sample(0:2,1000, replace=TRUE, prob=c(1/3, 1/2, 1/6))
Q <- sample(0:1,1000, replace=TRUE, prob=c(2/3, 1/3))
dfp <- new_quantile(P, "P")
dfq <- new_quantile(Q, "Q")
df_tq <- interpolate(dfp, dfq, "P")

}
