% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{logit}
\alias{logit}
\title{Logit transformation}
\usage{
logit(p)
}
\arguments{
\item{p}{a numeric value of probability, ranges between 0 and 1, exactly 0 and 1 not allowed}
}
\value{
a numeric value transformed to the real line
}
\description{
This function applies logit transformation for a given probability
}
\details{
The logit function transforms a probability within the range of 0 and 1 to the real line
}
