% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_plus.R
\name{logit_Louvain_clustering}
\alias{logit_Louvain_clustering}
\title{Louvain clustering using departure as data representation}
\usage{
logit_Louvain_clustering(
  scppp_obj,
  pdat = NULL,
  PCA = T,
  N = 15,
  pres = 0.8,
  tsne = F,
  umap = F
)
}
\arguments{
\item{pdat}{a matrix representated by model departure for each entry}

\item{PCA}{whether apply PCA before Louvain clustering, default is \code{TRUE}}

\item{N}{the number of pricipal components included for further clustering, default = 15}

\item{pres}{the resolution parameter in Louvain clustering, default = 0.8}

\item{test_set}{a UMI count data frame or matirx with cells as rows and genes as columns}
}
\value{
a list with the following elements:
\itemize{
\item{\code{sdata}}: {a Seurat object}
\item{\code{tsne_data}}: {a matrix containing t-SNE dimensionality reduction results,
with cells as rows, and first two t-SNE dimension as columns}
\item{\code{umap_data}}: {a matrix containing UMAP dimensionality reduction results,
with cells as rows, and first two UMAP dimension as columns}
\item{\code{sdata$seurat_clusters}}: {a vector with clustering index for each cell}
}
}
\description{
logit_Seurat_clustering returns a list with elements useful to check and compare cell clustering
}
\details{
This is a function used to get cell clustering using Louvain clustering implemented in Seurat package
}
\examples{

test_set <- matrix(rpois(500, 0.5), nrow = 10)
logit_Seurat_clustering(test_set)

}
\references{
\itemize{
\item Stuart, T. et al. Comprehensive integration of single-cell data. Cell 177, 1888 (2019)
}
}
