% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LouvainDepart.R
\name{LouvainDepart}
\alias{LouvainDepart}
\title{Louvain clustering using departure as data representation}
\usage{
LouvainDepart(
  data,
  pdat = NULL,
  PCA = T,
  N = 15,
  pres = 0.8,
  tsne = F,
  umap = F,
  ...
)
}
\arguments{
\item{data}{A UMI count matirx with genes as rows and cells as columns or an S3 object for class 'scppp'.}

\item{pdat}{A matrix used as input for cell clustering. If not specify, the departure matrix will be calculated within the function.}

\item{PCA}{A logic value specifying whether apply PCA before Louvain clustering, default is \code{TRUE}.}

\item{N}{A numeric value specifying the number of principal components included for further clustering (default 15).}

\item{pres}{A numeric value specifying the resolution parameter in Louvain clustering (default 0.8)}

\item{tsne}{A logic value specifying whether t-SNE dimension reduction should be applied for visualization.}

\item{umap}{A logic value specifying whether UMAP dimension reduction should be applied for visualization.}

\item{...}{not used.}
}
\value{
A list with the following elements:
\itemize{
\item{\code{sdata}}: {a Seurat object}
\item{\code{tsne_data}}: {a matrix containing t-SNE dimension reduction results,
with cells as rows, and first two t-SNE dimensions as columns; NULL if \code{tsne = FLASE}.}
\item{\code{umap_data}}: {a matrix containing UMAP dimension reduction results,
with cells as rows, and first two UMAP dimensions as columns; NULL if \code{tsne = FLASE}.}
\item{\code{res_clust}}: {a data frame contains two columns: names (cell names) and clusters (cluster label)}
}
}
\description{
This function returns a list with elements useful to check and compare cell clustering.
}
\details{
This is a function used to get cell clustering using Louvain clustering algorithm implemented in the Seurat package.
}
\examples{

test_set <- matrix(rpois(500, 0.5), nrow = 10)
logit_Seurat_clustering(test_set)

}
\references{
\insertRef{Seurat}{scpoisson}
}
