% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qq_env_small.R
\name{nboot_small}
\alias{nboot_small}
\title{Random sample generation function to generate sets of samples from theoretical Poisson distribution.}
\usage{
nboot_small(x, lambda, R)
}
\arguments{
\item{x}{A numeric vector of sampled data points to compare with theoretical Poisson.}

\item{lambda}{A numeric value specifying mean for theoretical Poisson distribution.}

\item{R}{A numeric value specifying the number of simulated sets.}
}
\value{
A data frame contains simulated data and corresponding simulation index.
}
\description{
This function returns a data frame with generated sets of samples and simulation index.
}
\details{
This is a function used to simulate a given number sets of samples from a theoretical Poisson distribution
that match input samples on sample size and sample mean (or theoretical Poisson parameter).
Plotting these as envelopes in Q-Q plot shows the variability in shapes we can expect when
sampling from the theoretical Poisson distribution.
}
\examples{
nboot_small(rpois(100, 3), 3, 200)

}
