% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qq_env_small.R
\name{new_quantile}
\alias{new_quantile}
\title{A more "continuous" approximation of quantiles of samples with a few integer case}
\usage{
new_quantile(data, sample)
}
\arguments{
\item{data}{A numeric vector of sampled data points.}

\item{sample}{A character string denotes which sample data points come from.}
}
\value{
A data frame contains the corresponding probability from cumulative distribution function (CDF), sample name, and corresponding respective quantiles.
}
\description{
This function returns a data frame including data points and corresponding quantile.
}
\details{
This is a function developed to get quantile for samples with only a few integer values.
Define both \eqn{p_{-1} = 0} and \eqn{q_{-1} = 0}.
Replace the point mass at each integer \eqn{z} by a bar on the interval \eqn{[z – \frac{1}{2}, z+ \frac{1}{2}]}
with height \eqn{P(X = z)}. This is a more "continuous" approximation of quantiles in this case.
}
