% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qq_env_small.R
\name{new_quantile_pois}
\alias{new_quantile_pois}
\title{A more "continuous" approximation of quantile of samples from theoretical Poisson distribution.}
\usage{
new_quantile_pois(data, lambda)
}
\arguments{
\item{data}{a numeric vector of sampled data points to comapre with theoretical Poisson.}

\item{lambda}{a numeric value for theoretical Poisson distribution parameter (equal to mean).}
}
\value{
a data frame. First column is quantile, second column is "theoretical",
and third column is correponsding quantiles from sample data.
}
\description{
new_quantile_pois returns a data frame including the probability
from cumulative distribution function (CDF) and corresponding quantiles.
}
\details{
This is a function developed to get corresponding quantiles from theoretical Poisson distribution.
The data points ranges from 0 to maximum value of sampled data used to compare with theoretical Poisson.
}
\examples{

P <- rpois(100, 3)
new_quantile_pois(P, 3)

}
