% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qq_env_small.R
\name{qq_interpolation}
\alias{qq_interpolation}
\title{Paired quantile after interpolation between two samples}
\usage{
qq_interpolation(dfp, dfq, sample1, sample2)
}
\arguments{
\item{dfp}{a data frame generated from function new_quantile() with first distribution.}

\item{dfq}{a data frame generated from function new_quantile() with second distribution.}

\item{sample1}{a character to denote sample name of first distribution.}

\item{sample2}{a character to denote sample name of second distribution.}
}
\value{
a data frame. First column is corresponding probability
from cumulative distribution function (CDF),
second column is corresponding quanitles from the first sample,
and third  column is correponsding quanitles from the second sample.
}
\description{
qq_interpolation returns a data frame with paired quantiles in two samples after interpolation.
}
\details{
This is a function for quantile interpolation of two samples.
For each unique quantile value that has original data
point in one sample but no corresponding original data point in another sample,
apply a linear interpolation. So the common quantile values after interpolation
should have unique points the same as unique quantile points from either sample.
}
\examples{

P <- sample(0:2,1000, replace=TRUE, prob=c(1/3, 1/2, 1/6))
Q <- sample(0:1,1000, replace=TRUE, prob=c(2/3, 1/3))
dfp <- new_quantile(P, "P")
dfq <- new_quantile(Q, "Q")
df_tq <- qq_interpolation(dfp, dfq, "P", "Q")

}
