% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qq_env_small.R
\name{qqplot_env_pois}
\alias{qqplot_env_pois}
\title{Q-Q plot comparing samples with a theoretical Poisson distribution}
\usage{
qqplot_env_pois(sample_data, lambda, envelope_size = 100, ...)
}
\arguments{
\item{sample_data}{a numeric vector of sample data points}

\item{lambda}{a numeric value for theoretical Poisson parameter}

\item{envelope_size}{a numeric value of size of envelope on Q-Q plot (default 100)}
}
\value{
a ggplot object
}
\description{
qqplot_small_pois returns a Q-Q plot using a more "continuous" approximation of quantiles.
}
\details{
This is a function for Q-Q envelope plot used to compare whether given sample data points come from the
theoretical Poisson distribution.  By simulating repeated samples of the same size from the candidate
theoretical distribution, and overlaying the envelope on the same figure, it provides a feeling of
understanding the natural variation from the theoretical distribution.
}
\examples{

qqplot_env_pois(rpois(200, 3), 3, 100)

}
