% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qq_env_small.R
\name{qqplot_small_test}
\alias{qqplot_small_test}
\title{Q-Q plot comparing two samples with small discrete counts}
\usage{
qqplot_small_test(P, Q, sample1, sample2)
}
\arguments{
\item{P}{A numeric vector from one sample.}

\item{Q}{A numeric vector from the other sample.}

\item{sample1}{A character to denote sample name of one distribution \code{P} generated from.}

\item{sample2}{A character to denote sample name of the other distribution \code{Q} generated from.}
}
\value{
A ggplot object. Q-Q plot with continuity correction. Quantiles from one sample on the horizontal axis and corresponding qunatiles
from the other sample on the vertical axis.
}
\description{
This function returns a ggplot object used to visualize quantiles comparing distributions of two samples.
}
\details{
This is a function for quantile-quantile plot comapring comparing samples from two discrete distributions
after \emph{continuity correction} and linear interpolation
}
\examples{

P <- sample(0:2,1000, replace=TRUE, prob=c(1/3, 1/2, 1/6))
Q <- sample(0:1,1000, replace=TRUE, prob=c(2/3, 1/3))
qqplot_small_test(P, Q, "P", "Q")

}
