% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scpoissonmodel-package.R
\docType{package}
\name{scpoissonmodel-package}
\alias{scpoissonmodel}
\alias{scpoissonmodel-package}
\title{scpoissonmodel: Single Cell Poisson Probability Paradigm}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Yue Pan \email{yue221@live.unc.edu}

Authors:
\itemize{
  \item Justin Landis \email{jtlandis314@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})
  \item Dirk Dittmer \email{dirk_dittmer@med.unc.edu}
  \item James S. Marron \email{marron@unc.edu}
  \item Di Wu \email{did@email.unc.edu}
}

}
\keyword{internal}
