% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Object_representation.R
\name{scppp}
\alias{scppp}
\title{Generate New scppp object}
\usage{
scppp(data, sample = c("columns", "rows"))
}
\arguments{
\item{data}{input data - Usually a matrix of counts}

\item{sample}{by rows or columns}
}
\description{
Generate New scppp object
}
