% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hclustDepart.R
\name{sigp}
\alias{sigp}
\title{Significance for first split using sigclust2}
\usage{
sigp(test_dat, minSize = 10, sim = 100)
}
\arguments{
\item{test_dat}{A UMI count data matrix with cells as rows and genes as columns.}

\item{minSize}{A numeric value specifying the minimal allowable cluster size (the number of cells for the smallest cluster, default 10).}

\item{sim}{A numeric value specifying the number of simulations during the Monte Carlo simulation procedure (default 100).}
}
\value{
A list with the following elements:
\itemize{
\item{\code{p}}: {p-value for the first split}
\item{\code{z}}: {z-score for the first split}
\item{\code{shc_result}}: {a \code{shc} S3-object as defined in sigclust2 package}
\item{\code{clust2}}: {a vector with group index for each cell}
\item{\code{clust_dat}}: {a matrix of data representation used as input for hierarchical clustering}
}
}
\description{
This function returns a list with elements mainly generated from sigclust2.
}
\details{
This is a function used to calculate the significance level of the first split from hierarchical clustering
based on euclidean distance and Ward's linkage.
}
\examples{

test_set <- matrix(rpois(500, 0.5), nrow = 10)
para_est_new(test_set)

}
\references{
\itemize{
\item Kimes, P. K., Liu, Y., Hayes, D. N. & Marron, J. S. Statistical Significance for Hierarchical
Clustering. Biometrics 73, 811–821 (2017).
\item Michael Linderman (2019). Rclusterpp: Linkable C++ Clustering. R package version 0.2.5.
https://github.com/nolanlab/Rclusterpp
}
}
