% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{theme_dirk}
\alias{theme_dirk}
\title{Dirk theme ggplots}
\usage{
theme_dirk(
  base_size = 22,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22,
  time_stamp = FALSE
)
}
\arguments{
\item{base_size}{base font size, givin in pts.}

\item{base_family}{base font family}

\item{base_line_size}{base size for line elements}

\item{base_rect_size}{base size for rect elements}

\item{time_stamp}{Logical value to indicate if the current
time should be added as a caption to the plot. Helpful for
versioning of plots.}
}
\value{
list that can be added to a ggplot object
}
\description{
This function generates ggplot object with theme elements that Dirk appreciates on his ggplots
}
\examples{

p <- ggplot(iris, aes(Sepal.Width, Sepal.Length, color = Species)) + geom_point()
p + theme_dirk()
p + labs(title = "Even Larger Than Default") + theme_dirk(30)
p + theme_dirk(time_stamp = TRUE)


}
