% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qq_env_small.R
\name{nboot_small}
\alias{nboot_small}
\title{Random sample generation function to generate sets of samples from theoretical Poisson distribution.}
\usage{
nboot_small(x, lambda, R)
}
\arguments{
\item{x}{a numeric vector of sampled data points to compare with theoretical Poisson.}

\item{lambda}{a numeric value for mean of theoretical Poisson.}

\item{R}{a numeric value for mean of theoretical Poisson.}
}
\value{
A data frame contains simulated data and corresponding simulation index.
Random sample generation function to generate sets of samples from theoretical Poisson distribution.

nboot_small returns a data frame with generate sets of samples and simulation index.

This is a function used to simulate a given number sets of samples from a theoretical Poisson distribution
that match input samples on sample size and sample mean (or theoretical Poisson parameter).
Plotting these as envelopes in Q-Q plot shows the variability in shapes we can expect when
sampling from the theoretical Poisson distribution.

a numeric vector of number of simulation sets that match input samples on sample size
and sample mean (or theoretical Poisson parameter).
}
\description{
This function returns a data frame with generated sets of samples and simulation index.
}
\details{
This is a function used to simulate a given number sets of samples from a theoretical Poisson distribution
that match input samples on sample size and sample mean (or theoretical Poisson parameter).
Plotting these as envelopes in Q-Q plot shows the variability in shapes we can expect when
sampling from the theoretical Poisson distribution.
}
