% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qq_env_small.R
\name{qqplot_small_test}
\alias{qqplot_small_test}
\title{Q-Q plot comparing two samples with small discrete counts}
\usage{
qqplot_small_test(P, Q, sample1, sample2)
}
\arguments{
\item{P}{a numeric vector from one sample}

\item{Q}{a numeric vector from the other sample}

\item{sample1}{a character to denote sample name of first distribution.}

\item{sample2}{a character to denote sample name of second distribution.}
}
\value{
a ggplot object with quantiles from one sample on the horizontal axis and corresponding qunatiles
from the other sample on the vertical axis.
}
\description{
qqplot_small_test returns a ggplot object visualize quantiles comapring distributions of two samples
}
\details{
This is a function for quantile-quantile plot comapring comparing samples from two discrete distributions
after \emph{continuity correction} and linear interpolation
}
\examples{

P <- sample(0:2,1000, replace=TRUE, prob=c(1/3, 1/2, 1/6))
Q <- sample(0:1,1000, replace=TRUE, prob=c(2/3, 1/3))
qqplot_small_test(P, Q, "P", "Q")

}
