% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scpoisson-package.R
\docType{package}
\name{scpoisson-package}
\alias{scpoisson}
\alias{scpoisson-package}
\title{scpoisson: Single Cell Poisson Probability Paradigm}
\description{
Useful to visualize the Poissoneity (an independent Poisson statistical framework, where each RNA measurement for each cell comes from its own independent Poisson distribution) of Unique Molecular Identifier (UMI) based single cell RNA sequencing (scRNA-seq) data, and explore cell clustering based on model departure as a novel data representation.
}
\author{
\strong{Maintainer}: Yue Pan \email{yuep027@gmail.com}

Authors:
\itemize{
  \item Justin Landis \email{jtlandis314@gmail.com} (\href{https://orcid.org/0000-0001-5501-4934}{ORCID})
  \item Dirk Dittmer \email{dirk_dittmer@med.unc.edu}
  \item James S. Marron \email{marron@unc.edu}
  \item Di Wu \email{did@email.unc.edu}
}

}
\keyword{internal}
